# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Auriga Simulation Laboratory API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Scenario {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** corridor | room | intersection | stairs | ramp | elevator | outdoor | mixed */
  environment_type: string;
  /** walking | blind | mobility_impaired | mixed */
  user_type: string;
  /** slow | normal | fast */
  movement_speed?: string;
  /** easy | medium | hard | adversarial */
  difficulty: string;
  /** none | low | medium | high | extreme */
  hazard_density: string;
  hazard_types?: string[];
  /** none | low | medium | high */
  sensor_noise_level?: string;
  /** bright | normal | dim | dark | flickering */
  lighting_condition?: string;
  /** @nullable */
  obstacle_count?: number | null;
  dynamic_obstacles?: boolean;
  created_at: string;
}

export interface ScenarioInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  environment_type: string;
  user_type: string;
  movement_speed?: string;
  difficulty: string;
  hazard_density: string;
  hazard_types?: string[];
  sensor_noise_level?: string;
  lighting_condition?: string;
  obstacle_count?: number;
  dynamic_obstacles?: boolean;
}

export interface ScenarioUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  environment_type?: string;
  user_type?: string;
  movement_speed?: string;
  difficulty?: string;
  hazard_density?: string;
  hazard_types?: string[];
  sensor_noise_level?: string;
  lighting_condition?: string;
  obstacle_count?: number;
  dynamic_obstacles?: boolean;
}

export interface Simulation {
  id: number;
  scenario_id: number;
  /** @nullable */
  scenario_name?: string | null;
  /** pending | running | completed | failed | cancelled */
  status: string;
  /** Number of synthetic test cases to run */
  scale: number;
  /** standard | adversarial | stress */
  mode?: string;
  /** @nullable */
  progress_pct?: number | null;
  /** @nullable */
  total_tests?: number | null;
  /** @nullable */
  passed_tests?: number | null;
  /** @nullable */
  failed_tests?: number | null;
  /** @nullable */
  success_rate?: number | null;
  /** @nullable */
  hazard_detection_rate?: number | null;
  /** @nullable */
  false_positive_rate?: number | null;
  /** @nullable */
  false_negative_rate?: number | null;
  /** @nullable */
  avg_latency_ms?: number | null;
  /** @nullable */
  decision_stability_score?: number | null;
  /** @nullable */
  started_at?: string | null;
  /** @nullable */
  completed_at?: string | null;
  created_at: string;
  /** @nullable */
  notes?: string | null;
}

export interface SimulationInput {
  scenario_id: number;
  /** @minimum 1 */
  scale: number;
  mode?: string;
  notes?: string;
}

export interface Benchmark {
  id: number;
  /** hazard_detection | guidance_quality | decision_stability | full_suite | adversarial */
  suite: string;
  name?: string;
  /** pending | running | completed | failed */
  status: string;
  /** @nullable */
  scenario_count?: number | null;
  /** @nullable */
  success_rate?: number | null;
  /** @nullable */
  hazard_detection_rate?: number | null;
  /** @nullable */
  false_positive_rate?: number | null;
  /** @nullable */
  false_negative_rate?: number | null;
  /** @nullable */
  decision_stability_score?: number | null;
  /** @nullable */
  avg_latency_ms?: number | null;
  /** @nullable */
  min_latency_ms?: number | null;
  /** @nullable */
  max_latency_ms?: number | null;
  /** @nullable */
  confidence_calibration_score?: number | null;
  /** @nullable */
  weaknesses_found?: number | null;
  created_at: string;
  /** @nullable */
  completed_at?: string | null;
}

export interface BenchmarkInput {
  suite: string;
  name?: string;
  scenario_count?: number;
}

export type BenchmarkAggregateBySuiteItem = {
  suite: string;
  count: number;
  avg_success_rate: number;
};

export interface BenchmarkAggregate {
  total_benchmarks: number;
  avg_success_rate: number;
  avg_hazard_detection_rate: number;
  avg_false_positive_rate: number;
  avg_false_negative_rate: number;
  avg_latency_ms: number;
  total_weaknesses_found: number;
  by_suite?: BenchmarkAggregateBySuiteItem[];
}

export interface Observation {
  id: number;
  /** simulation_failure | unexpected_behavior | research_debt | weakness | anomaly */
  type: string;
  /** low | medium | high | critical */
  severity: string;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  simulation_id?: number | null;
  /** @nullable */
  benchmark_id?: number | null;
  /** distance_engine | memory_engine | world_model | hazard_engine | guidance_engine | sensor | unknown */
  subsystem?: string;
  tags?: string[];
  resolved: boolean;
  /** @nullable */
  resolution_notes?: string | null;
  created_at: string;
  /** @nullable */
  resolved_at?: string | null;
}

export interface ObservationInput {
  type: string;
  severity: string;
  /** @minLength 1 */
  title: string;
  description?: string;
  simulation_id?: number;
  benchmark_id?: number;
  subsystem: string;
  tags?: string[];
}

export interface ObservationUpdate {
  resolved?: boolean;
  resolution_notes?: string;
  severity?: string;
  tags?: string[];
}

export interface DashboardSummary {
  total_scenarios: number;
  total_simulations: number;
  total_tests_run: number;
  overall_success_rate: number;
  overall_hazard_detection_rate: number;
  overall_false_positive_rate?: number;
  overall_false_negative_rate?: number;
  active_simulations: number;
  total_benchmarks: number;
  open_observations: number;
  critical_observations: number;
  total_weaknesses_found: number;
  avg_latency_ms?: number;
}

export interface ActivityEntry {
  id: number;
  /** simulation | benchmark | observation | scenario */
  entity_type: string;
  entity_id: number;
  /** @nullable */
  entity_name?: string | null;
  action: string;
  /** @nullable */
  metadata?: string | null;
  created_at: string;
}

export type ListScenariosParams = {
environment_type?: string;
user_type?: string;
difficulty?: string;
};

export type ListSimulationsParams = {
status?: string;
scenario_id?: number;
limit?: number;
};

export type ListBenchmarksParams = {
suite?: string;
limit?: number;
};

export type ListObservationsParams = {
type?: string;
severity?: string;
resolved?: boolean;
limit?: number;
};

export type GetDashboardActivityParams = {
limit?: number;
};