# lib/api-zod/src/generated/types/benchmark.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Auriga Simulation Laboratory API
 * OpenAPI spec version: 0.1.0
 */

export interface Benchmark {
  id: number;
  /** hazard_detection | guidance_quality | decision_stability | full_suite | adversarial */
  suite: string;
  name?: string;
  /** pending | running | completed | failed */
  status: string;
  /** @nullable */
  scenario_count?: number | null;
  /** @nullable */
  success_rate?: number | null;
  /** @nullable */
  hazard_detection_rate?: number | null;
  /** @nullable */
  false_positive_rate?: number | null;
  /** @nullable */
  false_negative_rate?: number | null;
  /** @nullable */
  decision_stability_score?: number | null;
  /** @nullable */
  avg_latency_ms?: number | null;
  /** @nullable */
  min_latency_ms?: number | null;
  /** @nullable */
  max_latency_ms?: number | null;
  /** @nullable */
  confidence_calibration_score?: number | null;
  /** @nullable */
  weaknesses_found?: number | null;
  created_at: string;
  /** @nullable */
  completed_at?: string | null;
}