# lib/api-zod/src/generated/types/observation.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Auriga Simulation Laboratory API
 * OpenAPI spec version: 0.1.0
 */

export interface Observation {
  id: number;
  /** simulation_failure | unexpected_behavior | research_debt | weakness | anomaly */
  type: string;
  /** low | medium | high | critical */
  severity: string;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  simulation_id?: number | null;
  /** @nullable */
  benchmark_id?: number | null;
  /** distance_engine | memory_engine | world_model | hazard_engine | guidance_engine | sensor | unknown */
  subsystem?: string;
  tags?: string[];
  resolved: boolean;
  /** @nullable */
  resolution_notes?: string | null;
  created_at: string;
  /** @nullable */
  resolved_at?: string | null;
}