# lib/api-zod/src/generated/types/observationUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Auriga Simulation Laboratory API
 * OpenAPI spec version: 0.1.0
 */

export interface ObservationUpdate {
  resolved?: boolean;
  resolution_notes?: string;
  severity?: string;
  tags?: string[];
}