# lib/api-zod/src/generated/types/scenarioUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Auriga Simulation Laboratory API
 * OpenAPI spec version: 0.1.0
 */

export interface ScenarioUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  environment_type?: string;
  user_type?: string;
  movement_speed?: string;
  difficulty?: string;
  hazard_density?: string;
  hazard_types?: string[];
  sensor_noise_level?: string;
  lighting_condition?: string;
  obstacle_count?: number;
  dynamic_obstacles?: boolean;
}